Address reviewer comments - commented-out and unused code.
